perf(no-ticket): keep the SDK and requests out of startup
Move create_requests_session to core/session.py, which reads the SDK
Configuration defaults only when the SDK is already imported. Defer the
SDK, session and saml imports in the decorators, the keyring provider
and the OIDC detectors to the code paths that use them.
