+++ cloudsmith_cli/core/credentials/models.py
@@ -3,9 +3,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
-import requests
+# requests costs ~30ms to import and is only used in annotations here.
+if TYPE_CHECKING:
+    import requests
 
 
 @dataclass
