+++ cloudsmith_cli/core/credentials/oidc/detectors/github_actions.py
@@ -15,7 +15,6 @@ from __future__ import annotations
 import os
 from urllib.parse import quote
 
-from ....rest import create_requests_session as create_session
 from .base import EnvironmentDetector
 
 DEFAULT_AUDIENCE = "cloudsmith"
@@ -41,6 +40,10 @@ class GitHubActionsDetector(EnvironmentDetector):
         audience = self.context.oidc_audience or DEFAULT_AUDIENCE
         separator = "&" if "?" in request_url else "?"
         url = f"{request_url}{separator}audience={quote(audience, safe='')}"
+
+        # The session module pulls in requests (~30ms). Import it here
+        # so that only a detector match pays that cost.
+        from ....session import create_requests_session as create_session
 
         session = self.context.session or create_session()
         try:
