+++ cloudsmith_cli/core/credentials/oidc/exchange.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING
 
 import requests
 
-from ...rest import create_requests_session as create_session
+from ...session import create_requests_session as create_session
 
 if TYPE_CHECKING:
     from ... import CredentialContext
