+++ cloudsmith_cli/core/credentials/providers/keyring_provider.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import logging
 
-from ....cli.saml import refresh_access_token
 from ....core import keyring
 from ..models import CredentialContext, CredentialResult
 from ..provider import CredentialProvider
@@ -34,6 +33,11 @@ class KeyringProvider(CredentialProvider):
                         "Session unavailable; skipping token refresh, using existing token"
                     )
                 else:
+                    # The saml module pulls in requests and the SDK
+                    # (~100ms). Import it here so that only the
+                    # token-refresh path pays that cost.
+                    from ....cli.saml import refresh_access_token
+
                     refresh_token = keyring.get_refresh_token(api_host)
                     new_access_token, new_refresh_token = refresh_access_token(
                         api_host,
